green: Returns_full_if_slots_are_full
